nova-act/nova-act/requirements.txt
cryptography
fire
install_playwright
jsonschema
pandas
pillow
playwright==1.48.0
pydantic
requests
retry
pytz
